refactor: adjust to latest BE response
