+++ src/graphql/getters/getSeasonBadges/index.ts
@@ -1,0 +1,2 @@
+export * from './getSeasonBadges';
+export * from './helpers';
